ann: move collect_features from features
